HIIT-023 refactor: 기존 ItTimeInfoExecutorManager 관련 클래스 service/manager로 패키지 이동
- ItTimeInfoExecutorManager를 다른 도메인에서도 사용하기 때문에 재사용성을 높이기 위해 service/manager 패키지로 이동
