+++ src/main/java/com/hiit/api/domain/service/manager/it/time/ItTimeInfo.java
@@ -1,0 +1,20 @@
+package com.hiit.api.domain.service.manager.it.time;
+
+import java.time.LocalTime;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@ToString
+@EqualsAndHashCode
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder(toBuilder = true)
+public class ItTimeInfo {
+	LocalTime startTime;
+	LocalTime endTime;
+}
